PCSkillsRelease/Tweet/requirements.txt
opencv_python==4.1.1.26
requests_oauthlib==1.3.0
tweepy==3.8.0
Pillow==6.2.1
